requirements_for_task1.txt
numpy==1.24.0
PyMuPDF==1.21.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pdfplumber==0.10.2
gdown==4.7.1
pytesseract==0.3.10
Pillow==9.6.1
